+++ CHANGELOG.md
@@ -45,6 +45,9 @@
 - Add PeaZip for an archive/compression utility.
 - Add BalenaEtcher as an image flashing utility.
 - Add support for resizing the root file system on NVMe drvies.
+- Add a first-time setup script.
+- Add support for NVIDIA graphics drivers.
+- Add support for Intel OpenGL graphics drivers on SteamOS.
 
 ## 2.2.0